add metrics

+++ test/test_util/tpch/workload.cpp
@@ -168,7 +168,7 @@ void Workload::Execute(int8_t worker_id, uint64_t execution_us_per_worker, uint6
 }
 
 uint64_t Workload::TimeQuery(int32_t query_ind, execution::vm::ExecutionMode mode, bool print_output) {
-  NOISEPAGE_ASSERT(query_ind < this->GetQueryNum() && 0 <= query_ind, "query plans index out of range");
+  NOISEPAGE_ASSERT(static_cast<uint32_t>(query_ind) < this->GetQueryNum() && 0 <= query_ind, "query plans index out of range");
   // Register to the metrics manager
   db_main_->GetMetricsManager()->RegisterThread();
   auto txn = txn_manager_->BeginTransaction();
